third_party/WebKit/LayoutTests/accessibility/aom-computed-accessible-node.html

<!DOCTYPE HTML>
<script src="../resources/gc.js"></script>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<!--

Accessibility Object Model
Explainer: https://github.com/WICG/aom/blob/master/explainer.md
Spec: https://wicg.github.io/aom/spec/

-->

<div id="button1" role="button">Click</div>

<script>

test(function(t) {
    assert_true(internals.runtimeFlags.accessibilityObjectModelEnabled);
}, "Make sure that Accessibility Object Model is enabled");

async_test(async function(test) {
    var element = document.getElementById("button1");
    var computedAXNode = await window.getComputedAccessibleNode(element)
    assert_true(computedAXNode != null);
    test.done();
}, 'Ensure that a non null value is returned from getComputedAccessibleNode');


async_test(async function(test) {
    var element = document.getElementById("button1");

    var computedAXNode1 = await window.getComputedAccessibleNode(element);
    var computedAXNode2 = await window.getComputedAccessibleNode(element);
    assert_true(computedAXNode1 === computedAXNode2);
    test.done();
}, "Multiple calls to getComputedAccessibleNode will return the same node.");

</script>

<div id="button2" role="button" aria-label="axButton">Click</div>

<script>

async_test(async function(test) {
    var button1 = document.getElementById("button1");
    var button1CAXNode = await window.getComputedAccessibleNode(button1);

    assert_equals(button1CAXNode.name, "Click");
    assert_equals(button1CAXNode.role, "button");

    button1.style.display = "none";

    var button2 = document.getElementById("button2");
    var button2CAXNode = await window.getComputedAccessibleNode(button2);

    assert_equals(button2CAXNode.name, "axButton");
    assert_equals(button2CAXNode.role, "button");

    // As button1 has no node in the accessibility tree anymore, assert that the
    // its previously retrieved computed accessible node has had its attributes
    // nullified.
    assert_equals(button1CAXNode.name, null);
    assert_equals(button1CAXNode.role, null);
    test.done();

}, "Deleting nodes from the accessibility tree will not cause a crash, and properties on any references to a deleted computed accessible node have been nullified.");

</script>
